requirements.txt
selenium==4.33.0
pytest==8.4.0
psycopg2-binary==2.9.10